Okay i lied
This is just the layout that avery and I want to use for the leadership
page, maybe put a border around the images, standardize their size, and
use better formatting to put the blurbs and he images side by side and
sized correctly. also these are not the final images we will be using.